Initialize all consoles in console-screen.kbd.sh, closes: #443606
In b9f2a…, the init script was changed to perform font/unicode mode
initialization on tty0 only when called in runlevel S.  It was planned to
let the script re-run later, when entering runlevels 1-5, to prepare the
remaining VTs; this was never implemented.  As there are enough font
changes going on during the bootup process (see #443645), simply drop the
RUNLEVEL = S case.  This should not generate any problems.