Rename “console-screen.kbd.sh” to “kbd”
AFAICT, there is no reason to use an overly complicated name for the
init script.  It seems to make much more sense to stick to the
conventions.

This commit also changes the “Provides:” line in the LSB header to
“kbd”, which will break some init scripts (those depending on
“console-screen”) when insserv is used.  The fix for this problem is a
virtual name (such as $console) which will have to be coordinated with
insserv, console-setup and others.