files with the same name installed in / and /usr
The package installs a symbolic link and a file with the same name in
both /bin/ and /usr/bin/, so it makes impossible to convert a system
to the everything-in-usr directories scheme.

This solves the problem by creating the link in postinst and only if it
is needed.

For more information about everything-in-usr please read
http://anonscm.debian.org/cgit/users/md/usrmerge.git/tree/debian/README.Debian

Closes: #767922