Drop kbd init script and /etc/kbd/* if unused
We leave the files in the hands of the local admin if they seem to be
used. We're working under the assumption that most people use
console-setup, but that is not a hard requirement on Debian so we don't
want to yank the carpet under the feet for potential users of the kbd
script.