Stop creating open (-> openvt) compatibility symlink
This compatibility is ancient and should not be needed anymore.
I doubt many people expect the open command to map to openvt (and it
would likely make more sense to use open for xdg-open at some point in
the future).
Fedora maintainer has been contacted and also thought it would be fine
to drop the compatibility symlink on their side. Help welcome in
identifying and contacting people in other distributions which also
ships the compat symlink.

Closes: #732796