Disable libkeymap
This means the keymap parts will instead be statically linked,
which avoids the problem of the udeb depending on a shared library
which isn't available in the udeb. (Also the kbd binary package should
not ship the shared objects, they should be in a separate package
really.)

In the future we might want to enable building the shared libkeymap
again, but before that we likely need to change debian/rules to do
a proper multi-flavour build for the udeb and the deb instead of
the current hand-crafted rules that duplicates part of the upstream
build system.

Closes: #969965